requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==2.2.6
Django==4.1.7
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-countries==7.5.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
mailjet-rest==1.3.4
MarkupSafe==2.1.2
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
whitenoise==6.3.0
